backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Session==0.4.0
psycopg2-binary==2.9.1
python-dotenv==0.19.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
alembic==1.13.1
twilio==8.10.0
APScheduler==3.10.4
bcrypt==3.2.0

